refactor(server): Move handler from server to handler
