+++ internal/server/server.go
@@ -1,7 +1,6 @@
 package server
 
 import (
-	"encoding/json"
 	"net/http"
 
 	"github.com/go-dummy/dummy/internal/config"
@@ -47,50 +46,4 @@ func (s *Server) Run() error {
 	}
 
 	return nil
-}
-
-// Handler -.
-func (s *Server) Handler(w http.ResponseWriter, r *http.Request) {
-	if setStatusCode(w, r.Header.Get("X-Set-Status-Code")) {
-		return
-	}
-
-	w.Header().Set("Content-Type", "application/json")
-
-	path := RemoveFragment(r.URL.Path)
-
-	response, ok := s.Handlers.Get(path, r.Method)
-	if ok {
-		w.WriteHeader(response.StatusCode)
-		resp := response.ExampleValue(r.Header.Get("X-Example"))
-
-		if nil == resp {
-			return
-		}
-
-		bytes, err := json.Marshal(resp)
-		if err != nil {
-			s.Logger.Error().Err(err).Msg("serialize response")
-		}
-
-		_, err = w.Write(bytes)
-		if err != nil {
-			s.Logger.Error().Err(err).Msg("write response")
-		}
-
-		return
-	}
-
-	w.WriteHeader(http.StatusNotFound)
-}
-
-func setStatusCode(w http.ResponseWriter, statusCode string) bool {
-	switch statusCode {
-	case "500":
-		w.WriteHeader(http.StatusInternalServerError)
-
-		return true
-	default:
-		return false
-	}
 }
